SkyPro/Lesson/lesson_12_2_test/pyproject.toml
[tool.poetry]
name = "12"
version = "0.1.0"
description = ""
authors = ["Javoprav <78508571+Javoprav@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
